requirements.txt
torch>=2.0.0
transformers>=4.30.0
google-genai>=0.3.0
accelerate>=0.20.0
sentencepiece>=0.1.99
bitsandbytes>=0.41.0
pandas>=1.5.0
langchain-community>=0.0.20
langchain-huggingface>=0.0.1
langchain-text-splitters>=0.0.1
wikipedia>=1.4.0
faiss-cpu>=1.7.4
pillow>=9.0.0
networkx>=3.0
matplotlib>=3.7.0
numpy>=1.24.0
tqdm>=4.65.0
tenacity>=8.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
